[Application] Fix code entry type enrichment
